+++ src/test/java/EstoqueTest.java
@@ -3,7 +3,6 @@ import Model.enums.Category;
 import Model.exceptions.ProdutoJaExisteException;
 import Model.exceptions.ProdutoNaoExisteException;
 import Model.services.Estoque;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -219,7 +218,7 @@ class EstoqueTest {
             while ((line = reader.readLine()) != null) {
                 String[] parts = line.split(",");
                 if (parts.length >= 6) {
-                    double id = Double.parseDouble(parts[0]);
+                    int id = Integer.parseInt(parts[0]);
                     String name = parts[1];
                     String manufacturer = parts[2];
                     Category category = Category.valueOf(parts[3]);
